chore: infra:db-main 구성 - infra 모듈 빈 구현체 주입 안되는 오류
