+++ commerce-domain/api/src/main/kotlin/com/example/domain/HealthApplication.kt
@@ -1,9 +1,9 @@
-package com.hanghae.health
+package com.example.domain
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = ["com.example"])
 class HealthApplication
 
 fun main(args: Array<String>) {
