+++ commerce-domain/api/src/main/kotlin/com/example/domain/controller/DemoController.kt
@@ -1,4 +1,4 @@
-package com.hanghae.health.controller
+package com.example.domain.controller
 
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
