+++ commerce-domain/api/src/main/kotlin/com/example/domain/service/UserRepository.kt
@@ -1,7 +1,6 @@
-package com.hanghae.health.service
+package com.example.domain.service
 
 interface UserRepository {
-
     fun save(user: User)
     fun read(id: Long): User?
 }
