+++ commerce-domain/api/src/test/kotlin/com/hanghae/health/service/FooServiceTest.kt
@@ -1,11 +1,3 @@
 package com.hanghae.health.service
 
-import org.junit.jupiter.api.Test
-
-class FooServiceTest {
-
-    @Test
-    fun foo() {
-        FooService().foo1()
-    }
-}
+class FooServiceTest
