+++ commerce-infra/db-main/src/main/kotlin/com/example/data/config/JpaConfiguration.kt
@@ -1,4 +1,4 @@
-package com.hanghae.storage.config
+package com.sample.data.config
 
 import jakarta.annotation.PostConstruct
 import org.springframework.boot.autoconfigure.domain.EntityScan
@@ -7,13 +7,13 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories
 import org.springframework.transaction.annotation.EnableTransactionManagement
 
 @Configuration
+@EntityScan(basePackages = ["com.sample.data.domain"])
 @EnableTransactionManagement
-@EntityScan(basePackages = ["com.hanghae.storage.domain"])
-@EnableJpaRepositories(basePackages = ["com.hanghae.storage.domain"])
-class MainJpaConfiguration {
+@EnableJpaRepositories(basePackages = ["com.sample.data.domain"])
+class JpaConfiguration {
 
     @PostConstruct
     fun init() {
-        println("MainJpaConfiguration init")
+        println("JpaConfiguration init")
     }
 }
