+++ commerce-infra/db-main/src/main/kotlin/com/example/data/domain/JpaUserRepository.kt
@@ -1,0 +1,5 @@
+package com.sample.data.domain
+
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface JpaUserRepository : JpaRepository<UserEntity, Long>
