+++ commerce-infra/db-main/src/main/kotlin/com/example/data/domain/UserEntityRepository.kt
@@ -1,0 +1,19 @@
+package com.sample.data.domain
+
+import com.example.domain.service.User
+import com.example.domain.service.UserRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+class UserEntityRepository(
+    private val jpaUserRepository: JpaUserRepository,
+) : UserRepository {
+
+    override fun save(user: User) {
+        TODO("Not yet implemented")
+    }
+
+    override fun read(id: Long): User? {
+        TODO("Not yet implemented")
+    }
+}
