+++ commerce-infra/db-main/src/main/resources/db-main.yml
@@ -1,4 +1,16 @@
 spring:
+  jpa:
+    show-sql: false
+    hibernate:
+      ddl-auto: none
+    properties:
+      hibernate:
+        format_sql: true
+        dialect: org.hibernate.dialect.MySQLDialect
+    database-platform: org.hibernate.dialect.MySQL5InnoDBDialect
+
+---
+spring:
   config:
     activate:
       on-profile: local
@@ -8,15 +20,7 @@ spring:
     username: root
     password: 1234qwer
   jpa:
-    show-sql: true
-    hibernate:
-      ddl-auto: none
-    properties:
-      hibernate:
-        format_sql: true
-        dialect: org.hibernate.dialect.MySQLDialect
-    open-in-view: false
-    database-platform: org.hibernate.dialect.MySQL5InnoDBDialect
+    show-sql: false
 
 logging:
   level:
@@ -43,11 +47,3 @@ spring:
     url: jdbc:mysql://database-1.czxx7hbyzohy.ap-northeast-2.rds.amazonaws.com:3306/test?useSSL=false&useUnicode=true&serverTimezone=Asia/Seoul&allowPublicKeyRetrieval=true
     username: root
     password: qwer1234!
-  jpa:
-    show-sql: false
-    hibernate:
-      ddl-auto: none
-    properties:
-      hibernate:
-        format_sql: false
-
