--- commerce-storage/db-main/src/main/kotlin/com/hanghae/storage/domain/user/UserJpaRepository.kt
@@ -1,5 +1,0 @@
-package com.hanghae.storage.domain.user
-
-import org.springframework.data.jpa.repository.JpaRepository
-
-internal interface UserJpaRepository : JpaRepository<UserEntity, Long>
